Refactor add_book action
